+++ src/environments/environment.ts
@@ -7,4 +7,5 @@ export const environment: Environment = {
     apiUrl: 'http://localhost:3050/api/v1',
     /** GiB para web-ifc MEMORY_LIMIT. Máx. usable ≈ 4 (se acota a uint32). Default lib: 2. */
     maxMemoryRender: 2,
+    urlSocket: 'http://localhost:3050', // Cambia la URL según tu configuración
 };